refactor: move constant delay into spec schedule
